changed allocator + more

+++ os/kernel/types/scalar.cpp
@@ -14,7 +14,7 @@ typedef signed long long int64_t;
 typedef unsigned long long uint64_t;
 
 typedef int32_t intptr_t;
-typedef uint32_t uintptr_t;
+typedef uint64_t uintptr_t;
 
 #pragma pack(push, 1)
 struct [[gnu::packed]] _co_uint24_t {
